change schema
